Extracted delivery dialog to separate module
It will be used in Launcher and CLI
